requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pyngrok>=7.0.0
